requirements.txt
mlx-whisper==0.4.3
numpy==2.4.3
pyobjc==12.1
rumps==0.4.0
sounddevice==0.5.5
